+++ source/ps/ProfileViewer.cpp
@@ -1,4 +1,4 @@
-/* Copyright (C) 2024 Wildfire Games.
+/* Copyright (C) 2025 Wildfire Games.
  * This file is part of 0 A.D.
  *
  * 0 A.D. is free software: you can redistribute it and/or modify
@@ -195,7 +195,7 @@ void CProfileViewer::RenderProfile(CCanvas2D& canvas)
 	for (size_t col = 0; col < columns.size(); ++col)
 	{
 		CStrW text = columns[col].title.FromUTF8();
-		int w, h;
+		float w, h;
 		font.CalculateStringSize(text.c_str(), w, h);
 
 		float x = colX;
@@ -228,7 +228,7 @@ void CProfileViewer::RenderProfile(CCanvas2D& canvas)
 		for (size_t col = 0; col < columns.size(); ++col)
 		{
 			CStrW text = table->GetCellText(row, col).FromUTF8();
-			int w, h;
+			float w, h;
 			font.CalculateStringSize(text.c_str(), w, h);
 
 			float x = rowColX;
